Use EntityType in Mob Constructors.

+++ src/MiNET/MiNET/Entities/Hostile/Blaze.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Blaze : HostileMob
 	{
-		public Blaze(Level level) : base((int) EntityType.Blaze, level)
+		public Blaze(Level level) : base(EntityType.Blaze, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
