+++ src/MiNET/MiNET/Entities/Hostile/Ghast.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Ghast : HostileMob
 	{
-		public Ghast(Level level) : base((int) EntityType.Ghast, level)
+		public Ghast(Level level) : base(EntityType.Ghast, level)
 		{
 			Width = Length = 4.0;
 			Height = 4.0;
