+++ src/MiNET/MiNET/Entities/Passive/Bat.cs
@@ -4,7 +4,7 @@ namespace MiNET.Entities.Passive
 {
 	public class Bat : PassiveMob
 	{
-		public Bat(Level level) : base((int) EntityType.Bat, level)
+		public Bat(Level level) : base(EntityType.Bat, level)
 		{
 			Width = Length = 0.5;
 			Height = 0.9;
