+++ src/MiNET/MiNET/Entities/Passive/Wolf.cs
@@ -5,7 +5,7 @@ namespace MiNET.Entities.Passive
 {
 	public class Wolf : PassiveMob
 	{
-		public Wolf(Level level) : base((int) EntityType.Wolf, level)
+		public Wolf(Level level) : base(EntityType.Wolf, level)
 		{
 			Width = Length = 0.6;
 			Height = 0.8;
